Make separate segmentation module
